A lot of functionality
Added Assets with their corresponding value which makes for easy loading in the game.
Created a method which loads the image of the card as per its number.
Created a command-line version of the game which allows for two players to face each other all though shifting the position of the cards is tedious.

Also created a basic version of the game against the computer. The basic logic of the game being intact made creating working version of the game really easy.