Made the command line game more convenient and other changes
Made the command line game more convenient.
Made a display file for the command tried to make functions more modular.
Added a set sorter, which automatically sorts the hand from largest to smallest.
This was a precursor to setting up the rest of the AI.
Removed a superfluous dependency of logic to display.