Added playable game for Mac, added a ≈Cache
Added a cache (created a file with all the functions already run) for the ranks of hands as the calculations wont change.
Created a way to hash sets of Cards then created a simpler way that is a bit reminicent of bitsets from c++ but a bit more simple.
Also created a multiple player game but it is super buggy and I reckon its my intial slow Hash function that might be to blame so I am rooting it out slowly from my code.