Changed the HandRank to make the result a float
Changed Handrank to a float which allowed me to do away with the bubble sorts and use the sort() function for the values resulting in a very sped up AI and game in general compared to before.
There is still an issue of the function sometimes loading and sometimes not.